+++ resources/column_names.md
@@ -1,0 +1,23 @@
+# Column Names and descriptions for King County Data Set
+
+- **id** - unique identified for a house
+- **dateDate** - house was sold
+- **pricePrice** - is prediction target
+- **bedroomsNumber** - # of bedrooms
+- **bathroomsNumber** - # of bathrooms
+- **sqft_livingsquare** - footage of the home
+- **sqft_lotsquare** - footage of the lot
+- **floorsTotal** - floors (levels) in house
+- **waterfront** - House which has a view to a waterfront
+- **view** - quality of view
+- **condition** - How good the condition is ( Overall )
+- **grade** - overall grade given to the housing unit, based on King County grading system
+- **sqft_above** - square footage of house apart from basement
+- **sqft_basement** - square footage of the basement
+- **yr_built** - Built Year
+- **yr_renovated** - Year when house was renovated
+- **zipcode** - zip
+- **lat** - Latitude coordinate
+- **long** - Longitude coordinate
+- **sqft_living15** - The square footage of interior housing living space for the nearest 15 neighbors
+- **sqft_lot15** - The square footage of the land lots of the nearest 15 neighbors
