+++ resources/starters/README-starter.md
@@ -1,0 +1,59 @@
+[![Shipping files](https://github.com/neuefische/ds-eda-project-template/actions/workflows/workflow-03.yml/badge.svg?branch=main&event=workflow_dispatch)](https://github.com/neuefische/ds-eda-project-template/actions/workflows/workflow-03.yml)
+# ds-project-template
+
+Template for creating ds simple projects
+
+## Requirements
+
+- pyenv
+- python==3.11.3
+
+## Setup
+
+ Install the virtual environment and the required packages by following commands:
+
+    ```BASH
+    pyenv local 3.11.3
+    python3 -m venv .venv
+    source .venv/bin/activate
+    pip install --upgrade pip
+    pip install -r requirements.txt
+    ```
+### **`WindowsOS`** type the following commands :
+
+
+- `Step_1:` Update Chocolatey and install Node by following commands:
+    ```sh
+    choco upgrade chocolatey
+    choco install nodejs
+    ```
+
+- `Step_2:` Install the virtual environment and the required packages by following commands.
+
+   For `PowerShell` CLI :
+
+    ```PowerShell
+    pyenv local 3.11.3
+    python -m venv .venv
+    .venv\Scripts\Activate.ps1
+    pip install --upgrade pip
+    pip install -r requirements.txt
+    ```
+
+    For `Git-Bash` CLI :
+  
+    ```BASH
+    pyenv local 3.11.3
+    python -m venv .venv
+    source .venv/Scripts/activate
+    pip install --upgrade pip
+    pip install -r requirements.txt
+    ```
+ 
+
+ **`Note:`**
+    If you encounter an error when trying to run `pip install --upgrade pip`, try using the following command:
+
+   ```Bash
+   python.exe -m pip install --upgrade pip
+   ```
