+++ resources/starters/requirements-starter.txt
@@ -1,0 +1,9 @@
+altair==5.3.0
+seaborn==0.13.2
+jupyterlab==4.0.1
+ipywidgets==8.0.6
+jupyterlab-dash==0.1.0a3
+python-dotenv==1.0.0
+psycopg2-binary==2.9.7
+SQLAlchemy==2.0.15
+missingno==0.5.2
